Export functionality for user-defined extensions
